requirements.txt.txt
fastapi
uvicorn
pillow
google
python-dotenv